Clicking cells works
Fixed the app-state bug -- was holding the whole response from the server in there
with body being string. Now I'm parsing the response and populating the app-state
during init with a proper map. This pretty much fixed the previous form of clicking
(without bakend) and also cleaned up the board-view component as well.