Board generation implemented
The board is now generated based on parameters, which we could either configure
via properties or a get request (not implemented).

Bug fixed: EDN data passed from server to client was wrapped in get-respose twice,
which made the data include the headers and body parts, and also pass the main
data as string. Now it just works.