Printing the state for debug
The only actual result of today's coding session is a number of functions that help me "see" the
board. Should be useful for debugging.

My algorithm approach was silly and didn't work. I was trying to be too smart. That's a no-go for
flood fill approach. But I have drafted a new approach on paper. I will continue soon with first
just doing the filling (without trying to be smart and applying additional logic).

P.S. I've removed all the stuff from the project.clj file. That stuff was done a long time ago - I don't know what it is now. I will be doing a proper dependency setup when I'll need it. Wanna take this project one step at a time to learn and understand... how the sausage is made.