Flooded the board successfully
Used flooding to get a "trail" from target cell.
Then used the trail to find if we can touch any borders with it. If we can means the target cell is not surrounded. This check will need to be updated to add different types of borders (friendly walls).

The code needs refactoring though: there are duplications. But I want to make it work first and make sure the logic is covered with tests and then see how I can make it better.