Second OMFG this month - not so simple test passed
Also this commit includes like a 1000% test coverage improvement. Had to finally do it to make sure I can refactor in the future and find the bugs that caused my "bigger" tests to fail. Some of the refactorings have already been done here as well - just a part of fixing faulty logic or misuse on FP.