Multiple logical bugs fixed
After adding another test that checks surrounding of the cells that surrounding a cell I've noticed a bunch of bugs in my code: hardcoded values covering state issues.
Now fixed and the complex recursive test is passing. The surround of russounding mechanism might change in the future after we can test it on in an actual gameplay and see if it makes sense or not.

For me it makes sense: even if you've surrounded a large area/group of enemies, you can still be surrounded by the enemy if the board allows. So you cannot just chill about it.