Drawing walls based on a graph
Decided that my walls will be represented as a graph (vector of lists flavor). That way it's easier to draw and construct since the data-type is a classic and easy to think about.
Currently the drawing is done but I do need to filter out diagonal walls where I can get away with using horizontal and vertical only. Already have the pattern figured out in my head just need to code it.

Also, switched to a more complex example for the drawing. Still looking ok.