Added walls calculation and drawing on each move
It behaves like a game now but the performance makes it impossible to play and a pain to test.
It seems that the performance of the algorithm is broken and is getting slower on each move (don't know why yet). And the rendering is quite slow as well. It looks like it re-renders the whole board instead of only the zots that changed, which is stupid and probably my fault. Must investigate further and learn stuff.