Some silly optimizations added.
The big optimization was - stop as soon as a direction touches a wall during the flood fill procedure. Not sure if it breaks anything though. The walls drawing mechanism was broken anyway just in a different way. At least with this optimization it's possible to manually test the game engine.

The second optimization is pretty much nothing, just something that kinda makes sense.

Added notes.md file to provide some insight into my state of mind regarding this project at the moment.