Implemented convex hull algorithm - simple surround works
The walls are built by a convex hull algorithm. It looks pretty good on screen but it does not support clusters of walls yet - all the walls blocks considered to be a single array and thus the hull is created for all of them. The next step will be to create clusters (vectors) of wall-cells and then just run each cluster through the convex hull algorithm. This means that the wall-ui part should also support vector of vectors instead of a single vector. But that should be straightforward.

I did not clean the unused code yet since I might need inspiration or reuse from the previous hand-made wall algorithm.