Switched from convex hull to my own implementation - outline claws.
The name of the algorithm I came up with is self-explanatory, right? :)
In reality the algorithm is based on gift-wrapping convex hull. But since I didn't need convex hull I had to mutate it into something that's best described as claws grabbing a cluster of points.
It's based on sorting by angles and then walking the vector one point at a time till you reach the end. Then do the same but use the other side (right hand vs left hand).

It's still dirty code so next time I'll cleanup and add tests.