Added unit tests for wall module + shape determination.
Started to cover wall module with unit tests. Adding different game situations and fixing bugs.

An important change: noticed that for my wall drawing algorithm to work I need to first determine the shape of the cluster of points. If the shape is vertical in structure I sort by Y first and start drawing with the lowest point. In other cases I sort by X first and start drawing with the left most point. Test-played the game a little and noticed that the drawing is still broken sometimes. I will try to figure out the minimal test case and address the issue.

Found an important bug and fixed it: fill flood should be able to fill cells that are taken by enemies but already surrounded. It's pretty much your teritory :)

Added more spec checks but not much.

Still need to address the biggest issue at the moment: sometimes the JS crashes when the game state is quite complex and someon makes the next move. The exception is "index out of bounds" when wall/find-next is getting called. So I shall start by covering this guy with tests and then try to figure out the issue. Maybe it's just an edge case for validation or maybe the bug is from the game engine itself - it gets into an invalid state somehow.