Change the vertical shape determination mechanism.
It seems to me that the shape of a cluster of wall cells should be based on the first cell and last cell coords after sorting. Only then the shape fits the algorithm that draws the walls.

Added another test case to cover the situation.
Still having another issue with JS crashing but it's a really exotic one.