Added a game spec.
This full game spec will allow me to have a reference for what the game state should contain when the server generates/updates it and sends back to client.

The spec is not final I'm sure but it's a start.

Already found and fixed a missing definition in a spec of :player slot in board. It was missing a possible :none value. Profit!