Implemented POST /game/:id flow of interceptors.
Based on the flow diagram I did in a notebook and presented to myself (oh, loneliness).
The flow is:
1. Check if the game exists.
2. Check if the move is valid.
3. Make the move (update db; return new game state).

The database is still in memory so for testing in postman I need to do get /game first to create a game in memory and then post moves to it. Works ok.