Added GET /game/:id flow
Used a new technique this time. I have a dispatcher interceptor that read the "accepted" header provider by the content negotiation interceptor. The dispatcher then decides which interceptor to add to the queue in order to render the response. I like it. Will do the same for the GET /game flow. This makes sure we have the same chain for different views.