Added an endless loop based.
The function was taken from Nolen's twit where he linked this gist: https://gist.github.com/swannodette/5882703

The loop is a naive implementation: send request; wait; repeat. But it's done in a go block so it's non-blocking. In the future I would like to switch to proper core.async usage and web sockets so I can push to the client and trigger and event.