Added Turn Text that is dynamic and as per design
The turn text will change color and text based on whose turn it is at the moment to give players an easy to see signal when to go and when to wait.
Also loading the correct font and using that.
Also removed the game title since it's not in the design.