Added score board boxes as per design
Used a stupid drawing with lines and circles to achieve the look for now.
Also cleaned up and removed some spacing between top and bottom.
I can tweak the css later but for now it's close to the real thing.

P.S. I've tried to increase the size of the board but that makes the game a bit slower than comfortable level of interaction. Will need to do performance improvements first if I want to provide a bigger board size.