Made the :red slot to be created with the game
Added more unit tests that cover game setup scenario to make sure I'm safe if I want to do refactoring of the slots/cookie stamping code. And I really want to do it cause it's a bit ugly.

Increased the delay between pulling from the server from 5 to 15 seconds because in real-life connectivity 5 seconds was too slow and made requests queue up in the browser eating up computer's mem and cpu. This will go away when I switch to web sockets in the future (I hope).