Wrote my own walls walker again - doesn't work
It works in some cases but there are still cases when the direction of the walk suddenly changes and it breaks my algorithm. I did some researching and found a nice concave hull paper. Will try it next.

Also played with interceptors a little bit to improve slotting. Still not happy. I may choose to rewrite interceptors in a seaparate branch when walls are done.