Implemented concave hull algorithm
The concave hull algorithm was based on a whitepaper I found (will credint in README).
But had to be modified a little bit due to an additional problem I have with my input data - a point in a cluster of points that can be used twice during the collection of hull. Someday I will write a proper article that explains this problem and the solution I've found.

I've added four tricky tests for this algorithm. These tests cover most of the known "problematic" cases I came across during the initial play-tests of the game.

Important lesson that I've had to learn in order to implement, test, debug this algorithm without any pain was: use the simplest model that you need in your functions. In my example I've switched from using models of game dots (x, y, surrouded?, player) to generic coordinates [x y]. This free my mind from reading through a lot of unneeded info and let me create test data and check the output even manually.

Also, don't fight clojure. A lot of algorithm from the CS world are based on collections with indeces. Clojure lets your work with data directly without using indeces. I was fighting Clojure on this during my first walls implementation but this time I've just used simple data directly and it was so much easier to reason about.

Me <3 Clojure (and ClojureScript of course).