Several minor changes - moving stuff around and clean-up
It seems that I didn't have a bug afterall. The problem came from facebook messenger which would visit the link that you've sent to get a preview of the page. That made facebook messenger the blue player so the actual second player becomes an observer every time. I will need to research the messenger's protocol but it seems to me that this can happen with other messengers as well. So I would probably switch from auto-join to a speicific action required from user to be the second player, like a blue button that says play instead of the blue score indicator.

Also moved reconciler to its own file to clean-up a bit.

Also started to send no caching headers with html response and added versioning to js and css files just in case. Currently the versioning is done manually. As many other things in this project that can be automated in the future.