Blue Play button added
A change in the game setup flow. Because of the way FB messenger works (and probably other messengers and crawlers as well) I was having problems with sharing a started game with another player. The messenger will visit the link to grab open-graph meta tags or whatever it falls back to is there are no open graph tags and become the second player. So the real second player on opening the link will always be an observer.

Fixed that by creating a button for the second player. Until the button is clicked the second player is an observer.

The button required other changes like putting :slots information into :game state as well. Which actually made the interceptors flow much nicer and also provided me the freedom of adding some UI changes as well:
- the red player can now see if the blue palyer joined the game or not - the style of the blue score indicator will change;
- the board becomes "inactive" (just looks opaque) when it's not your turn (this I'm still thinking about).

Tests had to be updated to cover the new flow.