Switched from in-memory database as object to the actual datomic db
The DB interceptor is now saving data to datomic db.
Reading from DB is done as well but the current game model is a bit tricky to convert the db model to because of non-flat structure. I might need to finally refactor the game model to be a flat map of maps. That way I can do a simple map over the data to convert to and from db schema.

I've commented out all the service_test.clj file since I will need to find a proper way to mock the datomic db. It was simple to do with in-memory object database I should be able to do something similar for datomic.

Anyway, this commit will result in a non-working game but it has a bunch of working code nonetheless.