One big dirty commit - db works
Datomic finally persists the game data.
However, the commit has all the debugging stuff and inline pieces of code that I was playing with.
Also it's slow. Super slow. Slow as a high priority bug slow. I will need to debug, read more about datomic and improve this.
The commit is only a stepping stone to make sure I have it. Something that works but not supposed to be seen by good people of... anywhere!