Added memoization to max-x and max-y functions
Since there's no chance that the size of the board changes during game I can safely memoize these two functions since they will always return the same value and the value now is calculated. It used to be a const but I want to make the engine more flexible so I always calculate the size of the board. But once the calculation is done we can cache it.