Added AWS property file samples and a first version of run book
The Run book is the important part of the commit. It includes all the commands and steps I've used to configure datomic and run the game connecting to an AWS transactor.

Also this commit means that I've setup the AWS datomic system (storage + transactor) and the game is currently sending the data and reading the data from a remote transactor. So the games are now persistent.