Made Turn message color consistent
The color of the turn message used to be changing depending on which player's turn it is. So it was constantly switching between blue and red. It was confusing to understand after awhile which color is player playing as. Now the color of the message is always the same as the player's color. So player can get used to the message being red if she is red and blue is she is blue.

Also updated the versions up.