Added notification box for Red Player to tell her to send the link to Blue player
The notification box is a bit ugly but it's by design right now. My thinking is: the user will not like to look at the ugly yellowish message so whe can either close the game or do what the message says to get rid of it. We'll see how it works in real like.