Added tab notification when tab is not active and it's your turn again
This is like 100% side-effects! So I've jammed them into reconciler flows.
One is in the callback function that's called after getting the data from server: it compares player's cookie with the turn in the response and checks if the tab is hidden then changes the title to have * in it.
The other function resets the tab's title when you make a move. I've placed into :post remote query flow thing of om.next.

A bit icky but for starters it's good and I think this feature is a must have for casual gamers.