Only applying the datomic schema on getting the db-connection
Since the connection fetch happens once it's a good place to apply the schema.
It's easy to do now with integrant, cause it can be just a part of setting up the system.
So as soon as we build the db-connection and before we build the interceptors we apply the schema.

Hopefuly improves the DB part's load a bit.