Switched from :pre assert so s/valid? in a check for valid cell
The unit test started to fail and I've noticed that I'm using assertions to protect myself from invalid moves, which is not good cause assertions are not running on prod anymore (but for some reason are finally running in tests).
Changed to an explicit s/valid? check for spec's compliance.