Multiple performance improvements
Switched trail to be a set.
Switched visited to be a set.
Both changes let me remove my own api on checking if things are there or not and also stop converting data into vectors where not needed.

Also added some destructuring in a couple of methods instead of repeated map reads.

Updated the test to check for sets and removed the old stuff.