Fix for safari hover issue
Noticed that on Safari on Macs the zot's hover effect is unstable. Probably due to a difference in rendering circles and their strokes.

Fixed the bug by using a white stroke on white background with a big width that covers most of the zot that's now big and allows for a stable hover effect. Easier to show than explain :)